Changed all doubles into floats.

+++ trunk/quake/src/generic/server.h
@@ -39,10 +39,10 @@ typedef struct
 	qboolean	paused;
 	qboolean	loadgame;			// handle connections specially
 
-	double		time;
+	float		time;
 	
 	int			lastcheck;			// used by PF_checkclient
-	double		lastchecktime;
+	float		lastchecktime;
 	
 	char		name[64];			// map name
 #ifdef QUAKE2
@@ -83,7 +83,7 @@ typedef struct client_s
 	qboolean		privileged;			// can execute any host command
 	qboolean		sendsignon;			// only valid before spawned
 
-	double			last_message;		// reliable messages must be sent
+	float			last_message;		// reliable messages must be sent
 										// periodically
 
 	struct qsocket_s *netconnection;	// communications handle
@@ -209,7 +209,7 @@ extern	client_t	*host_client;
 
 extern	jmp_buf 	host_abortserver;
 
-extern	double		host_time;
+extern	float		host_time;
 
 extern	edict_t		*sv_player;
